Added editprductpage, function to calculate the  expiration date, first step with secure storage

+++ lib/product_class.dart
@@ -6,7 +6,7 @@ class Product {
   final double amount;
   final String currency;
   final String store;
-  final double lengthWarranty;
+  final int lengthWarranty;
 
   Product({
     required this.id,
